Move GetAllConfigFields(), add unit test
